+++ src/components/link-box/index.ts
@@ -7,4 +7,5 @@ export {
   type LinkBoxLinkProps,
   type LinkBoxRootElement,
   type LinkBoxRootProps,
+  type LinkBoxVariant,
 } from "./LinkBox.js";